+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Updated to peakrdl-rust dependency crate v0.2.0
+- `Access` is now an associated type of the `Register` trait
+- `MemEntry` and `MemEntryIter` are now generic over the memory type instead of access and endianness.
+
 ## [0.6.2] - 2026-03-13
 
 ### Fixed