+++ crates/peakrdl-rust/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to the `peakrdl-rust` crate will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Changed
+
+- `Access` is now an associated type of the `Register` trait
+- `T: Register` generic of the `Reg` type was renamed to `R`
+
 ## [0.1.1] - 2026-03-13
 
 ### Fixed